+++ CHANGES.rst
@@ -100,6 +100,9 @@ Bug Fixes
 
 - ``astropy.io.ascii``
 
+  - Fix a segfault in the fast C parser when one of the column headers
+    is empty [#3545].
+
 - ``astropy.io.fits``
 
 - ``astropy.io.misc``