+++ CHANGES.rst
@@ -212,6 +212,21 @@ Miscellaneous
   - The ERFA wrappers are now written directly in the Python/C API
     rather than using Cython, for greater performance. [#3521]
 
+1.0.1 (unreleased)
+------------------
+
+Bug Fixes
+^^^^^^^^^
+
+- ``astropy.io.ascii``
+
+  - Fixed support for reading inf and nan values with the fast reader in
+    Windows.  Also fixed in the case of using ``use_fast_converter=True``
+    with the fast reader. [#3525]
+
+  - Fixed use of mmap in the fast reader on Windows. [#3525]
+
+
 1.0 (2015-02-18)
 ----------------